// ClinicaAPI/ClinicaApp/src/app/models/Tables/TableData.ts
export interface TableData {
  Foto: string;
  Ficha: string;
  Id?: number;
  Nome: string;
  Nascimento: any;
  Area: string;
  selecionada: boolean;
  SaiSoz: string;
  Idade: string;
  Desde: any;
  Proxses?: any;
  TelefoneFixo?: string;
  Celular?: string;
  TelCom?: string;
  NomeMae? : string;
  NomePai? : string;
  DtIncl? : any;
  RestrMae? : string;
  RestrPai? : string;
  RespFin? : string;
  Identidade? : string;
  Endereco? : string;
  Email? : string;
  Ativo? : boolean;
}